+++ tests/Functional/DoctrineOrm/DoctrineOrmRulerTest.php
@@ -104,6 +104,30 @@ class DoctrineOrmRulerTest extends TestCase
         $this->ruler->apply($qb, 'foo.bar', []);
     }
 
+    #[Test]
+    public function shouldKeepParametersAlreadySetOnQueryBuilder(): void
+    {
+        $qb = (new QueryBuilder($this->entityManager))
+            ->from(Product::class, 'products')
+            ->select('products')
+            ->andWhere('products.published = :published')
+            ->setParameter('published', true);
+
+        $this->ruler->apply($qb, 'price > :price', ['price' => 100]);
+
+        $parameters = [];
+
+        /** @var Parameter $parameter */
+        foreach ($qb->getParameters() as $parameter) {
+            $parameters[$parameter->getName()] = $parameter->getValue();
+        }
+
+        self::assertEquals([
+            'published' => true,
+            'price'     => 100,
+        ], $parameters);
+    }
+
     public static function provideDataForApply(): array
     {
         return [